+++ mock/control-api/src/api/mod.rs
@@ -58,6 +58,10 @@ pub struct AppContext {
 
 /// Run REST [Control API] server mock.
 ///
+/// # Panics
+///
+/// If the given `args` don't contain an expected `medea_addr` value.
+///
 /// [Control API]: https://tinyurl.com/yxsqplq7
 pub async fn run(
     args: &ArgMatches<'static>,
@@ -151,7 +155,7 @@ macro_rules! gen_request_macro {
 /// # Errors
 ///
 /// Errors if unable to send message to [`GrpcCallbackServer`] actor.
-#[allow(clippy::needless_pass_by_value)]
+#[allow(clippy::missing_panics_doc, clippy::needless_pass_by_value)]
 pub async fn get_callbacks(
     state: Data<AppContext>,
 ) -> Result<HttpResponse, ()> {
@@ -295,12 +299,13 @@ pub struct ErrorResponse {
     pub element: String,
 }
 
-impl Into<ErrorResponse> for proto::Error {
-    fn into(self) -> ErrorResponse {
-        ErrorResponse {
-            code: self.code,
-            text: self.text,
-            element: self.element,
+impl From<proto::Error> for ErrorResponse {
+    #[inline]
+    fn from(e: proto::Error) -> Self {
+        Self {
+            code: e.code,
+            text: e.text,
+            element: e.element,
         }
     }
 }
@@ -335,29 +340,29 @@ pub struct Response {
     pub error: Option<ErrorResponse>,
 }
 
-/// Macro which implements [`Into`] [`HttpResponse`] for all
-/// `control-api-mock` responses.
+/// Macro that implements [`From`] `control-api-mock` responses for
+/// [`HttpResponse`].
 ///
 /// Implementation will check existence of `error` and if it exists then
 /// [`HttpResponse`] will be `BadRequest` with this struct as response in
 /// otherwise `Ok` with this struct as response.
-macro_rules! impl_into_http_response {
+macro_rules! impl_from_for_http_response {
     ($resp:tt) => {
-        impl Into<HttpResponse> for $resp {
-            fn into(self) -> HttpResponse {
-                if self.error.is_some() {
-                    HttpResponse::BadRequest().json(self)
+        impl From<$resp> for HttpResponse {
+            fn from(resp: $resp) -> Self {
+                if resp.error.is_some() {
+                    Self::BadRequest().json(resp)
                 } else {
-                    HttpResponse::Ok().json(self)
+                    Self::Ok().json(resp)
                 }
             }
         }
     };
 }
 
-impl_into_http_response!(CreateResponse);
-impl_into_http_response!(Response);
-impl_into_http_response!(SingleGetResponse);
+impl_from_for_http_response!(CreateResponse);
+impl_from_for_http_response!(Response);
+impl_from_for_http_response!(SingleGetResponse);
 
 impl From<proto::Response> for Response {
     fn from(resp: proto::Response) -> Self {
@@ -395,6 +400,11 @@ pub enum Element {
 }
 
 impl Element {
+    /// Converts this [`Element`] into an appropriate [`proto::room::Element`].
+    ///
+    /// # Panics
+    ///
+    /// If a conversion for such an [`Element`] isn't implemented yet.
     #[must_use]
     pub fn into_proto(self, id: String) -> proto::room::Element {
         let el = match self {