+++ src/conf/server.rs
@@ -1,6 +1,6 @@
 //! Settings for application servers.
 
-use std::net::{IpAddr, Ipv4Addr, SocketAddr, ToSocketAddrs as _};
+use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 
 use derive_more::{Display, From};
 use serde::{Deserialize, Serialize};
@@ -64,15 +64,11 @@ pub struct ClientApiHttpServer {
 }
 
 impl ClientApiHttpServer {
-    /// Builds [`SocketAddr`] from `bind_ip` and `bind_port`.
+    /// Builds a [`SocketAddr`] from `bind_ip` and `bind_port`.
     #[inline]
     #[must_use]
     pub fn bind_addr(&self) -> SocketAddr {
-        (self.bind_ip, self.bind_port)
-            .to_socket_addrs()
-            .unwrap()
-            .next()
-            .unwrap()
+        (self.bind_ip, self.bind_port).into()
     }
 }