+++ src/signalling/elements/endpoints/webrtc/publish_endpoint.rs
@@ -2,11 +2,11 @@
 
 use std::{
     cell::RefCell,
-    collections::HashSet,
+    collections::{HashMap, HashSet},
     rc::{Rc, Weak},
 };
 
-use medea_client_api_proto::PeerId;
+use medea_client_api_proto::{PeerId, TrackId};
 use medea_control_api_proto::grpc::api as proto;
 
 use crate::{
@@ -25,6 +25,10 @@ use super::play_endpoint::WebRtcPlayEndpoint;
 struct WebRtcPublishEndpointInner {
     /// ID of this [`WebRtcPublishEndpoint`].
     id: Id,
+
+    /// [`TrackId`]s of the [`MediaTrack`]s related to this
+    /// [`WebRtcPublishEndpoint`].
+    tracks_ids: HashMap<PeerId, Vec<TrackId>>,
 
     /// P2P connection mode for this [`WebRtcPublishEndpoint`].
     p2p: P2pMode,
@@ -63,7 +67,7 @@ impl Drop for WebRtcPublishEndpointInner {
             .filter_map(WeakWebRtcPlayEndpoint::safe_upgrade)
         {
             if let Some(receiver_owner) = receiver.weak_owner().safe_upgrade() {
-                receiver_owner.remove_sink(&receiver.id())
+                drop(receiver_owner.remove_sink(&receiver.id()))
             }
         }
     }
@@ -137,6 +141,7 @@ impl WebRtcPublishEndpoint {
             audio_settings,
             video_settings,
             peer_ids: HashSet::new(),
+            tracks_ids: HashMap::new(),
         })))
     }
 
@@ -190,13 +195,6 @@ impl WebRtcPublishEndpoint {
         self.0.borrow_mut().reset()
     }
 
-    /// Removes [`PeerId`] from this [`WebRtcPublishEndpoint`]'s `peer_ids`.
-    #[allow(clippy::trivially_copy_pass_by_ref)]
-    #[inline]
-    pub fn remove_peer_id(&self, peer_id: &PeerId) {
-        self.0.borrow_mut().remove_peer_id(peer_id)
-    }
-
     /// Removes all [`PeerId`]s related to this [`WebRtcPublishEndpoint`].
     #[inline]
     pub fn remove_peer_ids(&self, peer_ids: &[PeerId]) {
@@ -233,6 +231,28 @@ impl WebRtcPublishEndpoint {
     #[must_use]
     pub fn is_force_relayed(&self) -> bool {
         self.0.borrow().is_force_relayed
+    }
+
+    /// Adds [`TrackId`] of the [`MediaTrack`] related to this
+    /// [`WebRtcPublishEndpoint`].
+    ///
+    /// [`MediaTrack`]: crate::media::track::MediaTrack
+    #[inline]
+    pub fn add_track_id(&self, peer_id: PeerId, track_id: TrackId) {
+        let mut inner = self.0.borrow_mut();
+        inner.tracks_ids.entry(peer_id).or_default().push(track_id);
+    }
+
+    /// Returns [`TrackId`]s of the related to this [`WebRtcPublishEndpoint`]
+    /// [`MediaTrack`]s from the [`Peer`] with a provided [`PeerId`].
+    ///
+    /// [`MediaTrack`]: crate::media::track::MediaTrack
+    /// [`Peer`]: crate::media::peer::Peer
+    #[inline]
+    #[must_use]
+    pub fn get_tracks_ids_by_peer_id(&self, peer_id: PeerId) -> Vec<TrackId> {
+        let inner = self.0.borrow();
+        inner.tracks_ids.get(&peer_id).cloned().unwrap_or_default()
     }
 
     /// Returns `true` if `on_start` or `on_stop` callback is set.
@@ -302,33 +322,35 @@ impl WeakWebRtcPublishEndpoint {
     }
 }
 
-impl Into<proto::WebRtcPublishEndpoint> for WebRtcPublishEndpoint {
-    fn into(self) -> proto::WebRtcPublishEndpoint {
-        let p2p: proto::web_rtc_publish_endpoint::P2p = self.p2p().into();
-        proto::WebRtcPublishEndpoint {
+impl From<WebRtcPublishEndpoint> for proto::WebRtcPublishEndpoint {
+    fn from(endpoint: WebRtcPublishEndpoint) -> Self {
+        let p2p: proto::web_rtc_publish_endpoint::P2p = endpoint.p2p().into();
+        Self {
             p2p: p2p as i32,
-            id: self.id().to_string(),
-            force_relay: self.is_force_relayed(),
-            audio_settings: Some(self.audio_settings().into()),
-            video_settings: Some(self.video_settings().into()),
+            id: endpoint.id().to_string(),
+            force_relay: endpoint.is_force_relayed(),
+            audio_settings: Some(endpoint.audio_settings().into()),
+            video_settings: Some(endpoint.video_settings().into()),
             on_stop: String::new(),
             on_start: String::new(),
         }
     }
 }
 
-impl Into<proto::member::Element> for WebRtcPublishEndpoint {
-    fn into(self) -> proto::member::Element {
-        proto::member::Element {
-            el: Some(proto::member::element::El::WebrtcPub(self.into())),
+impl From<WebRtcPublishEndpoint> for proto::member::Element {
+    #[inline]
+    fn from(endpoint: WebRtcPublishEndpoint) -> Self {
+        Self {
+            el: Some(proto::member::element::El::WebrtcPub(endpoint.into())),
         }
     }
 }
 
-impl Into<proto::Element> for WebRtcPublishEndpoint {
-    fn into(self) -> proto::Element {
-        proto::Element {
-            el: Some(proto::element::El::WebrtcPub(self.into())),
+impl From<WebRtcPublishEndpoint> for proto::Element {
+    #[inline]
+    fn from(endpoint: WebRtcPublishEndpoint) -> Self {
+        Self {
+            el: Some(proto::element::El::WebrtcPub(endpoint.into())),
         }
     }
 }