+++ tests/e2e/object/tracks_store.rs
@@ -3,6 +3,8 @@
 
 use std::marker::PhantomData;
 
+use derive_more::Display;
+
 use crate::{
     browser::Statement,
     object::{
@@ -14,6 +16,23 @@ use crate::{
 };
 
 use super::Error;
+
+/// Representation of a [MediaStreamTrackState][1].
+///
+/// [1]: https://w3.org/TR/mediacapture-streams/#dom-mediastreamtrackstate
+#[derive(Clone, Copy, Debug, Eq, Display, PartialEq)]
+pub enum MediaStreamTrackState {
+    /// Track is active (the track's underlying media source is making a
+    /// best-effort attempt to provide data in real time).
+    #[display(fmt = "live")]
+    Live,
+
+    /// Track has ended (the track's underlying media source is no longer
+    /// providing data, and will never provide more data for this track). Once
+    /// a track enters this state, it never exits it.
+    #[display(fmt = "ended")]
+    Ended,
+}
 
 /// Shortcut for a [`TracksStore`] of [`LocalTrack`]s.
 pub type LocalTracksStore = TracksStore<LocalTrack>;
@@ -181,19 +200,25 @@ impl<T> Object<TracksStore<T>> {
 
     /// Checks whether all local `Track`s from this store are in the `ended`
     /// `readyState`.
-    pub async fn is_all_tracks_ended(&self) -> Result<bool, Error> {
+    pub async fn all_tracks_have_ready_state(
+        &self,
+        ready_state: MediaStreamTrackState,
+    ) -> Result<bool, Error> {
         self.execute(Statement::new(
             // language=JavaScript
-            r#"
-                async (store) => {
-                    for (track of store.tracks) {
-                        if (track.track.get_track().readyState != 'ended') {
-                            return false;
-                        }
-                    }
-                    return true;
-                }
-            "#,
+            &format!(
+                r#"
+                    async (store) => {{
+                        for (track of store.tracks) {{
+                            if (track.track.get_track().readyState !== '{}') {{
+                                return false;
+                            }}
+                        }}
+                        return true;
+                    }}
+                "#,
+                ready_state,
+            ),
             [],
         ))
         .await?