+++ jason/CHANGELOG.md
@@ -6,6 +6,26 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [0.2.0] · 2021-??-?? · To-be-done
+[0.2.0]: /../../tree/medea-jason-0.2.0/jason
+
+[Diff](/../../compare/medea-jason-0.2.0-rc.1...medea-jason-0.2.0) | [Milestone](/../../milestone/2) | [Roadmap](/../../issues/27)
+
+First, consider [0.2.0-rc.1] changes.
+
+### Added
+
+- Signalling:
+    - Handling of RPC events:
+        - `PeerUpdated` with `PeerUpdate::Removed` ([#109]).
+- Library API:
+    - `RemoteMediaTrack.on_stopped` callback that is called when `RemoteMediaTrack` is stopped ([#109]).
+
+[#109]: /../../pull/109
+
+
+
+
 ## [0.2.0-rc.1] · 2021-02-01
 [0.2.0-rc.1]: /../../tree/medea-jason-0.2.0-rc.1/jason